#38_#39 Repositories long->doubrle

+++ src/main/java/sda/fitapp/entity/Ingredient.java
@@ -13,10 +13,10 @@ public class Ingredient {
     @Column(name = "id")
     private long id;
     private String name;
-    private long kcal100g;
-    private long gramsOfProteins;
-    private long gramsOfCarbs;
-    private long gramsOfFats;
+    private double kcal100g;
+    private double gramsOfProteins;
+    private double gramsOfCarbs;
+    private double gramsOfFats;
     private boolean proteinProduct;
     private boolean carbProduct;
     private boolean fatProduct;
